requirements.txt
opencv-python 
torch 
transformers 
Pillow